1. merge conversation_instr into conversation;2. add supply of deepseek-v3 template
